The login functionality has been created

+++ AuthOn.Application/Configurations/TokenConfiguration.cs
@@ -3,6 +3,6 @@
     public class TokenConfiguration
     {
         public string Key { get; set; } = string.Empty;
-        public int ExpiresInHours { get; set; }
+        public double ExpiresInHours { get; set; }
     }
 }
